Module scc::hash_cache

HashCache is a concurrent and asynchronous 32-way associative cache backed by HashMap.

Structs

HashCache

Scalable concurrent 32-way associative cache backed by HashMap.

OccupiedEntry

OccupiedEntry is a view into an occupied cache entry in a HashCache.

VacantEntry

VacantEntry is a view into a vacant cache entry in a HashCache.

Enums

Entry

Entry represents a single cache entry in a HashCache.

Constants

DEFAULT_MAXIMUM_CAPACITY

The default maximum capacity of a HashCache is 256.

Type Aliases

EvictedEntry

EvictedEntry is a type alias for Option<(K, V)>.