scc::hash_cache

Struct HashCache

pub struct HashCache<K, V, H = RandomState>
where
    H: BuildHasher,
{ /* private fields */ }

Scalable concurrent 32-way associative cache backed by HashMap.

HashCache is a concurrent 32-way associative cache based on the HashMap implementation. HashCache does not keep track of the least recently used entry in the entire cache. Instead, each bucket maintains a doubly linked list of occupied entries, updated on access to entries to keep track of the least recently used entries within the bucket. Therefore, entries can be evicted before the cache is full.

HashCache and HashMap share the same runtime characteristics, except that each entry in a HashCache additionally uses 2-byte space for a doubly linked list and a HashCache starts evicting least recently used entries if the bucket is full instead of allocating a linked list of entries.

Unwind safety

HashCache is impervious to out-of-memory errors and panics in user-specified code under one condition: H::Hasher::hash, K::drop and V::drop must not panic.

Implementations

impl<K, V, H> HashCache<K, V, H>
where H: BuildHasher,

pub const fn with_hasher(build_hasher: H) -> Self

Creates an empty HashCache with the given BuildHasher.

The maximum capacity is set to DEFAULT_MAXIMUM_CAPACITY.

Examples

use scc::HashCache;
use std::collections::hash_map::RandomState;

let hashcache: HashCache<u64, u32, RandomState> = HashCache::with_hasher(RandomState::new());

pub fn with_capacity_and_hasher( minimum_capacity: usize, maximum_capacity: usize, build_hasher: H, ) -> Self

Creates an empty HashCache with the specified capacity and BuildHasher.

The actual capacity is equal to or greater than the specified capacity.

Examples

use scc::HashCache;
use std::collections::hash_map::RandomState;

let hashcache: HashCache<u64, u32, RandomState> =
    HashCache::with_capacity_and_hasher(1000, 2000, RandomState::new());

let result = hashcache.capacity();
assert_eq!(result, 1024);

impl<K, V, H> HashCache<K, V, H>
where K: Eq + Hash, H: BuildHasher,

pub async fn entry_async(&self, key: K) -> Entry<'_, K, V, H>

Gets the entry associated with the given key in the map for in-place manipulation.

Examples

use scc::HashCache;

let hashcache: HashCache<char, u32> = HashCache::default();

let future_entry = hashcache.entry_async('b');

pub fn entry_sync(&self, key: K) -> Entry<'_, K, V, H>

Gets the entry associated with the given key in the map for in-place manipulation.

Examples

use scc::HashCache;

let hashcache: HashCache<char, u32> = HashCache::default();

for ch in "a short treatise on fungi".chars() {
    hashcache.entry_sync(ch).and_modify(|counter| *counter += 1).or_put(1);
}

assert_eq!(*hashcache.get_sync(&'s').unwrap().get(), 2);
assert_eq!(*hashcache.get_sync(&'t').unwrap().get(), 3);
assert!(hashcache.get_sync(&'y').is_none());

pub fn try_entry(&self, key: K) -> Option<Entry<'_, K, V, H>>

Tries to get the entry associated with the given key in the map for in-place manipulation.

Returns None if the entry could not be locked.

Examples

use scc::HashCache;

let hashcache: HashCache<usize, usize> = HashCache::default();

assert!(hashcache.put_sync(0, 1).is_ok());
assert!(hashcache.try_entry(0).is_some());

pub async fn put_async( &self, key: K, val: V, ) -> Result<EvictedEntry<K, V>, (K, V)>

Puts a key-value pair into the HashCache.

Returns Some if an entry was evicted for the new key-value pair.

Note

If the key exists, the value is not updated.

Errors

Returns an error containing the supplied key-value pair if the key exists.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();
let future_put = hashcache.put_async(11, 17);

pub fn put_sync(&self, key: K, val: V) -> Result<EvictedEntry<K, V>, (K, V)>

Puts a key-value pair into the HashCache.

Returns Some if an entry was evicted for the new key-value pair.

Note

If the key exists, the value is not updated.

Errors

Returns an error containing the supplied key-value pair if the key exists.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert_eq!(hashcache.put_sync(1, 1).unwrap_err(), (1, 1));

pub async fn replace_async(&self, key: K) -> ReplaceResult<'_, K, V, H>

Adds a key to the HashCache, replacing the existing key, if any, that is equal to the given one.

Examples

use std::cmp::{Eq, PartialEq};
use std::hash::{Hash, Hasher};

use scc::HashCache;
use scc::hash_cache::ReplaceResult;

#[derive(Debug)]
struct MaybeEqual(u64, u64);

impl Eq for MaybeEqual {}

impl Hash for MaybeEqual {
    fn hash<H: Hasher>(&self, state: &mut H) {
        // Do not read `self.1`.
        self.0.hash(state);
    }
}

impl PartialEq for MaybeEqual {
    fn eq(&self, other: &Self) -> bool {
        // Do not compare `self.1`.
        self.0 == other.0
    }
}

let hashcache: HashCache<MaybeEqual, usize> = HashCache::default();

async {
    let ReplaceResult::NotReplaced(v) = hashcache.replace_async(MaybeEqual(11, 7)).await else {
        unreachable!();
    };
    drop(v.put_entry(17));
    let ReplaceResult::Replaced(_, k) = hashcache.replace_async(MaybeEqual(11, 11)).await else {
        unreachable!();
    };
    assert_eq!(k.1, 7);
};

pub fn replace_sync(&self, key: K) -> ReplaceResult<'_, K, V, H>

Adds a key to the HashCache, replacing the existing key, if any, that is equal to the given one.

Examples

use std::cmp::{Eq, PartialEq};
use std::hash::{Hash, Hasher};

use scc::HashCache;
use scc::hash_cache::ReplaceResult;

#[derive(Debug)]
struct MaybeEqual(u64, u64);

impl Eq for MaybeEqual {}

impl Hash for MaybeEqual {
    fn hash<H: Hasher>(&self, state: &mut H) {
        // Do not read `self.1`.
        self.0.hash(state);
    }
}

impl PartialEq for MaybeEqual {
    fn eq(&self, other: &Self) -> bool {
        // Do not compare `self.1`.
        self.0 == other.0
    }
}

let hashcache: HashCache<MaybeEqual, usize> = HashCache::default();

let ReplaceResult::NotReplaced(v) = hashcache.replace_sync(MaybeEqual(11, 7)) else {
    unreachable!();
};
drop(v.put_entry(17));
let ReplaceResult::Replaced(_, k) = hashcache.replace_sync(MaybeEqual(11, 11)) else {
    unreachable!();
};
assert_eq!(k.1, 7);

pub async fn remove_async<Q>(&self, key: &Q) -> Option<(K, V)>
where Q: Equivalent<K> + Hash + ?Sized,

Removes a key-value pair if the key exists.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();
let future_put = hashcache.put_async(11, 17);
let future_remove = hashcache.remove_async(&11);

pub fn remove_sync<Q>(&self, key: &Q) -> Option<(K, V)>
where Q: Equivalent<K> + Hash + ?Sized,

Removes a key-value pair if the key exists.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.remove_sync(&1).is_none());
assert!(hashcache.put_sync(1, 0).is_ok());
assert_eq!(hashcache.remove_sync(&1).unwrap(), (1, 0));

pub async fn get_async<Q>(&self, key: &Q) -> Option<OccupiedEntry<'_, K, V, H>>
where Q: Equivalent<K> + Hash + ?Sized,

Gets an OccupiedEntry corresponding to the key for in-place modification.

OccupiedEntry exclusively owns the entry, preventing others from gaining access to it: use read_async if read-only access is sufficient.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();
let future_put = hashcache.put_async(11, 17);
let future_get = hashcache.get_async(&11);

pub fn get_sync<Q>(&self, key: &Q) -> Option<OccupiedEntry<'_, K, V, H>>
where Q: Equivalent<K> + Hash + ?Sized,

Gets an OccupiedEntry corresponding to the key for in-place modification.

OccupiedEntry exclusively owns the entry, preventing others from gaining access to it: use read_sync if read-only access is sufficient.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.get_sync(&1).is_none());
assert!(hashcache.put_sync(1, 10).is_ok());
assert_eq!(*hashcache.get_sync(&1).unwrap().get(), 10);

*hashcache.get_sync(&1).unwrap() = 11;
assert_eq!(*hashcache.get_sync(&1).unwrap(), 11);

pub async fn read_async<Q, R, F: FnOnce(&K, &V) -> R>( &self, key: &Q, reader: F, ) -> Option<R>
where Q: Equivalent<K> + Hash + ?Sized,

Reads a key-value pair.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();
let future_put = hashcache.put_async(11, 17);
let future_read = hashcache.read_async(&11, |_, v| *v);

pub fn read_sync<Q, R, F: FnOnce(&K, &V) -> R>( &self, key: &Q, reader: F, ) -> Option<R>
where Q: Equivalent<K> + Hash + ?Sized,

Reads a key-value pair.

Returns None if the key does not exist.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.read_sync(&1, |_, v| *v).is_none());
assert!(hashcache.put_sync(1, 10).is_ok());
assert_eq!(hashcache.read_sync(&1, |_, v| *v).unwrap(), 10);

pub async fn contains_async<Q>(&self, key: &Q) -> bool
where Q: Equivalent<K> + Hash + ?Sized,

Returns true if the HashCache contains a value for the specified key.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

let future_contains = hashcache.contains_async(&1);

pub fn contains_sync<Q>(&self, key: &Q) -> bool
where Q: Equivalent<K> + Hash + ?Sized,

Returns true if the HashCache contains a value for the specified key.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(!hashcache.contains_sync(&1));
assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.contains_sync(&1));

pub async fn remove_if_async<Q, F: FnOnce(&mut V) -> bool>( &self, key: &Q, condition: F, ) -> Option<(K, V)>
where Q: Equivalent<K> + Hash + ?Sized,

Removes a key-value pair if the key exists and the given condition is met.

Returns None if the key does not exist or the condition was not met.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();
let future_put = hashcache.put_async(11, 17);
let future_remove = hashcache.remove_if_async(&11, |_| true);

pub fn remove_if_sync<Q, F: FnOnce(&mut V) -> bool>( &self, key: &Q, condition: F, ) -> Option<(K, V)>
where Q: Equivalent<K> + Hash + ?Sized,

Removes a key-value pair if the key exists and the given condition is met.

Returns None if the key does not exist or the condition was not met.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.remove_if_sync(&1, |v| { *v += 1; false }).is_none());
assert_eq!(hashcache.remove_if_sync(&1, |v| *v == 1).unwrap(), (1, 1));

pub async fn iter_async<F: FnMut(&K, &V) -> bool>(&self, f: F) -> bool

Iterates over entries asynchronously for reading.

Stops iterating when the closure returns false, and this method also returns false.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u64> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());

async {
    let result = hashcache.iter_async(|k, v| {
        false
    }).await;
    assert!(!result);
};

pub fn iter_sync<F: FnMut(&K, &V) -> bool>(&self, f: F) -> bool

Iterates over entries synchronously for reading.

Stops iterating when the closure returns false, and this method also returns false.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u64> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.put_sync(2, 1).is_ok());

let mut acc = 0_u64;
let result = hashcache.iter_sync(|k, v| {
    acc += *k;
    acc += *v;
    true
});

assert!(result);
assert_eq!(acc, 4);

pub async fn iter_mut_async<F: FnMut(ConsumableEntry<'_, '_, K, V>) -> bool>( &self, f: F, ) -> bool

Iterates over entries asynchronously for modification.

Stops iterating when the closure returns false, and this method also returns false.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.put_sync(2, 1).is_ok());

async {
    let result = hashcache.iter_mut_async(|entry| {
        if entry.0 == 1 {
            entry.consume();
            return false;
        }
        true
    }).await;

    assert!(!result);
    assert_eq!(hashcache.len(), 1);
};

pub fn iter_mut_sync<F: FnMut(ConsumableEntry<'_, '_, K, V>) -> bool>( &self, f: F, ) -> bool

Iterates over entries synchronously for modification.

Stops iterating when the closure returns false, and this method also returns false.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.put_sync(2, 1).is_ok());
assert!(hashcache.put_sync(3, 2).is_ok());

let result = hashcache.iter_mut_sync(|entry| {
    if entry.0 == 1 {
        entry.consume();
        return false;
    }
    true
});

assert!(!result);
assert!(!hashcache.contains_sync(&1));
assert_eq!(hashcache.len(), 2);

pub async fn retain_async<F: FnMut(&K, &mut V) -> bool>(&self, pred: F)

Retains the entries specified by the predicate.

This method allows the predicate closure to modify the value field.

Entries that have existed since the invocation of the method are guaranteed to be visited if they are not removed, however the same entry can be visited more than once if the HashCache gets resized by another thread.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

let future_put = hashcache.put_async(1, 0);
let future_retain = hashcache.retain_async(|k, v| *k == 1);

pub fn retain_sync<F: FnMut(&K, &mut V) -> bool>(&self, pred: F)

Retains the entries specified by the predicate.

This method allows the predicate closure to modify the value field.

Entries that have existed since the invocation of the method are guaranteed to be visited if they are not removed, however the same entry can be visited more than once if the HashCache gets resized by another thread.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert!(hashcache.put_sync(2, 1).is_ok());
assert!(hashcache.put_sync(3, 2).is_ok());

hashcache.retain_sync(|k, v| *k == 1 && *v == 0);

assert!(hashcache.contains_sync(&1));
assert!(!hashcache.contains_sync(&2));
assert!(!hashcache.contains_sync(&3));

pub async fn clear_async(&self)

Clears the HashCache by removing all key-value pairs.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

let future_put = hashcache.put_async(1, 0);
let future_clear = hashcache.clear_async();

pub fn clear_sync(&self)

Clears the HashCache by removing all key-value pairs.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
hashcache.clear_sync();

assert!(!hashcache.contains_sync(&1));

pub fn len(&self) -> usize

Returns the number of entries in the HashCache.

It reads the entire metadata area of the bucket array to calculate the number of valid entries, making its time complexity O(N). Furthermore, it may overcount entries if an old bucket array has yet to be dropped.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.put_sync(1, 0).is_ok());
assert_eq!(hashcache.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the HashCache is empty.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert!(hashcache.is_empty());
assert!(hashcache.put_sync(1, 0).is_ok());
assert!(!hashcache.is_empty());

pub fn capacity(&self) -> usize

Returns the capacity of the HashCache.

Examples

use scc::HashCache;

let hashcache_default: HashCache<u64, u32> = HashCache::default();
assert_eq!(hashcache_default.capacity(), 0);

let hashcache: HashCache<u64, u32> = HashCache::with_capacity(1000, 2000);
assert_eq!(hashcache.capacity(), 1024);

pub fn capacity_range(&self) -> RangeInclusive<usize>

Returns the current capacity range of the HashCache.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::default();

assert_eq!(hashcache.capacity_range(), 0..=256);

impl<K, V> HashCache<K, V, RandomState>

pub fn new() -> Self

Creates an empty default HashCache.

The maximum capacity is set to DEFAULT_MAXIMUM_CAPACITY.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::new();

let result = hashcache.capacity();
assert_eq!(result, 0);

pub fn with_capacity(minimum_capacity: usize, maximum_capacity: usize) -> Self

Creates an empty HashCache with the specified capacity.

The supplied minimum and maximum capacity values are adjusted to power-of-two values equal to or larger than the provided values and 64 with one exception; if 0 is specified as the minimum capacity, the minimum capacity of the HashCache becomes 0.

Examples

use scc::HashCache;

let hashcache: HashCache<u64, u32> = HashCache::with_capacity(1000, 2000);

let result = hashcache.capacity();
assert_eq!(result, 1024);

let hashcache: HashCache<u64, u32> = HashCache::with_capacity(0, 0);
let result = hashcache.capacity_range();
assert_eq!(result, 0..=64);