scc

Module hash_map

HashMap is a concurrent hash map.

Structs

ConsumableEntry

ConsumableEntry is a view into an occupied entry in a HashMap when iterating over entries in it.

HashMap

Scalable concurrent hash map.

OccupiedEntry

OccupiedEntry is a view into an occupied entry in a HashMap.

Reserve

Reserve keeps the capacity of the associated HashMap higher than a certain level.

VacantEntry

VacantEntry is a view into a vacant entry in a HashMap.

Enums

Entry

Entry represents a single entry in a HashMap.

ReplaceResult

ReplaceResult is the result type of the HashMap::replace_async and HashMap::replace_sync methods.