Struct scc::LinkedEntry

pub struct LinkedEntry<T> { /* private fields */ }

Entry stores an instance of T and a link to the next entry.

Implementations

impl<T> Entry<T>

pub unsafe fn take_inner(&mut self) -> T

Extracts the inner instance of T.

Safety

This method has to be called at most once per Entry, and the caller needs to make sure that the Entry is not accessed via LinkedList methods.

Examples

use scc::Stack;

let stack: Stack<usize> = Stack::default();

stack.push(37);

let mut entry = stack.pop().unwrap();
let pushed = unsafe { entry.get_mut().unwrap().take_inner() };
assert_eq!(pushed, 37);